// src/types/SearchType.ts
export enum SearchEnum {
  SORT = 'sort',
  ITEMS = 'items',
  QUERY = 'query',
  PAGE = 'page',
}

export enum SortEnum {
  Default = 'No sort',
  Newest = 'Newest',
  Oldest = 'Oldest',
  Expensivest = 'Expensivest',
  Cheapest = 'Cheapest',
}

export enum ItemsSortEnum {
  All = 'All',
  Twelve = '12 items',
  TwentyFour = '24 items',
  ThirtySix = '36 items',
  FortyEight = '48 items',
}

export enum ItemsValueEnum {
  Twelve = 12,
  TwentyFour = 24,
  ThirtySix = 36,
  FortyEight = 48,
}